Make Grid block a functional component
