+++ src/blocks/container/css/main.js
@@ -76,6 +76,7 @@ export default class MainCSS extends Component {
 			fontSizeUnit,
 			textTransform,
 			shapeDividers,
+			gridId,
 		} = attributes;
 
 		let containerWidthPreview = containerWidth;
