+++ src/blocks/container/edit.js
@@ -19,7 +19,6 @@ import DesktopCSS from './css/desktop.js';
 import TabletCSS from './css/tablet.js';
 import TabletOnlyCSS from './css/tablet-only.js';
 import MobileCSS from './css/mobile.js';
-import getAllUniqueIds from '../../utils/get-all-unique-ids';
 import getResponsivePlaceholder from '../../utils/get-responsive-placeholder';
 import hasNumericValue from '../../utils/has-numeric-value';
 import isBlockVersionLessThan from '../../utils/check-block-version';
@@ -92,16 +91,6 @@ class GenerateBlockContainer extends Component {
 	}
 
 	componentDidMount() {
-		// Generate a unique ID if none exists or if the same ID exists on this page.
-		const allBlocks = wp.data.select( 'core/block-editor' ).getBlocks();
-		const uniqueIds = getAllUniqueIds( allBlocks, [], this.props.clientId );
-
-		if ( ! this.props.attributes.uniqueId || uniqueIds.includes( this.props.attributes.uniqueId ) ) {
-			this.props.setAttributes( {
-				uniqueId: this.props.clientId.substr( 2, 9 ).replace( '-', '' ),
-			} );
-		}
-
 		const thisBlock = document.getElementById( 'block-' + this.props.clientId );
 
 		if ( thisBlock && 'full' === this.props.attributes.align ) {
