+++ src/blocks/grid/css/main.js
@@ -30,7 +30,7 @@ export default class MainCSS extends Component {
 			'margin-right': '-' + ( horizontalGap / 2 ) + 'px',
 		} ];
 
-		cssObj[ '.gb-grid-wrapper-' + uniqueId + ' > .block-editor-inner-blocks > .block-editor-block-list__layout > .wp-block' ] = [ {
+		cssObj[ '.gb-grid-wrapper-' + uniqueId + ' > .block-editor-inner-blocks > .block-editor-block-list__layout' ] = [ {
 			'padding-left': ( horizontalGap / 2 ) + 'px',
 			'padding-right': ( horizontalGap / 2 ) + 'px',
 			'margin-bottom': valueWithUnit( verticalGap, 'px' ),
