+++ src/hoc/index.js
@@ -1,0 +1,7 @@
+import withUniqueId from './withUniqueId';
+import withGridLegacyMigration from './withGridLegacyMigration';
+
+export {
+	withUniqueId,
+	withGridLegacyMigration,
+};
