+++ src/hooks/index.js
@@ -1,0 +1,7 @@
+import useDeviceType from './useDeviceType';
+import useInnerBlocksCount from './useInnerBlocksCount'
+
+export {
+	useDeviceType,
+	useInnerBlocksCount,
+};
