+++ src/utils/get-editor-blocks.js
@@ -1,0 +1,1 @@
+export default () => wp.data.select( 'core/block-editor' ).getBlocks();
